refactor: move gurbaninow into separate service
